requirements.txt
Flask==1.1.1
Pillow==7.0.0
psycopg2==2.8.4
requests==2.22.0
urllib3==1.25.8
uWSGI==2.0.18
